+++ eSTOL_Training_Tool/eSTOL_Training_Tool_Core/changelog.md
@@ -1,3 +1,8 @@
+v1.4.5
+
+- added check update button
+- added manual update option
+
 v1.4.4
 
 - added plane config Savage Norden